rel to #2012: refactor GCWebAPI before adding log edit: clean up API access
